### 2018-01-16 #532 @JoshKarpel
Fixed some errors in `paper.bib` that were causing formatting problems in the PDF (commit is on master branch).
