### 2018-01-17 #532 @kyleniemeyer
Hi @JoshKarpel, thanks for your submission. Do you have a suggested reviewer or two that we can start from? I usually try to get one suggested and one non-specialist from the JOSS reviewer pool.
