### 2018-01-17 #532 @JoshKarpel
Sounds like a good plan @kyleniemeyer. I took a look at the reviewer pool and found several reviewers who might be interested: @nicoguaro, @jarvist, @leios, and @iljah. They're all marked for Python and expressed some interest in numerical methods/computational physics (which is where I'm coming from) in their topic areas.
