### 2018-01-22 #542 @whedon
[REVIEW]: IDESolver: a general purpose integro-differential equation solver
**Submitting author:** @JoshKarpel (<a href="http://orcid.org/0000-0001-5968-9373">Josh Karpel</a>)
**Repository:** <a href="https://github.com/JoshKarpel/idesolver" target ="_blank">https://github.com/JoshKarpel/idesolver</a>
**Version:** v1.0.1
