### 2018-01-22 #532 @kyleniemeyer
OK @JoshKarpel @iljah @arghdos we'll continue the review over in #542
