### 2018-01-22 #542 @kyleniemeyer
OK @iljah @arghdos please add your review comments here— @iljah you should fill out the checkbox, and @arghdos you should make sure you agree that everything is filled/not filled (JOSS only officially supports one reviewer at the moment).
