### 2018-01-22 f377917
Create CONTRIBUTING.md
