+++ CONTRIBUTING.md
@@ -1,0 +1,15 @@
+# Contributing to IDESolver
+
+## Reporting Issues
+
+Report issues via the GitHub issue tracker.
+Be as specific as possible about the problem you're encountering, and include all relevant details (a [minimal working example](https://en.wikipedia.org/wiki/Minimal_Working_Example), the full traceback of the error or description of how the output and the desired output differ, Python version, IDESolver version, etc.)
+
+## Contributing Code
+
+We are generally open to GitHub pull requests.
+However, IDESolver is built to solve a very specific problem.
+We won't accept pull requests that expand the project beyond the core idea: solving IDEs numerically via iterative relaxation algorithms.
+
+* Follow [pep8](https://www.python.org/dev/peps/pep-0008/) when possible.
+* Extensions to functionality should generally be presented as optional additions/replacements to existing options.
