### 2018-01-22 93a1b4c
Added JOSS badge; links to issues, code of conduct, and contribution guidelines to readme.
