### 2018-01-22 #542 @JoshKarpel
I've added a code of conduct and contributor guidelines, along with links to them from the readme, on the master branch.
