### 2018-01-23 #542 @iljah
The second link in paper pdf is broken, probably because the doi is given as https... in paper.bib to which another https... is prepended.
