### 2018-01-23 JoshKarpel/idesolver#3 @iljah
Need root for `pip install idesolver`
I would be nice to mention in readme that the given pip command requires root and/or also provide a version for local installation: `pip install --user idesolver`
