### 2018-01-23 #542 @iljah
A question about documentation for @kyleniemeyer: reviewer instructions state e.g. "A high-level overview of this documentation should be included in a README file (or equivalent)." but [the readme file](https://github.com/JoshKarpel/idesolver/blob/master/README.rst) mostly refers to documentation elsewhere. The referred documentation is also in html format, is that ok for joss?

@JoshKarpel according to [reviewer instructions](https://joss.theoj.org/about#reviewer_guidelines) the readme should include the following items, could you please add them to readme or to the html docs?

- [ ] A statement of need
- [x] Example usage
- [x] API documentation
- [ ] Tests
- [x] Community guidelines
