### 2018-01-23 JoshKarpel/idesolver#4 @iljah
Invalid syntax in idesolver.py
After doing a `pip3 install --user idesolver` and trying to import idesolver in python3 I get:
```
Python 3.5.4 (default, Oct  9 2017, 12:07:29)
[GCC 6.4.1 20170727 (Red Hat 6.4.1-1)] on linux
>>> from idesolver import IDESolver
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File ".../.local/lib/python3.5/site-packages/idesolver/__init__.py", line 1, in <module>
    from .idesolver import *
  File ".../.local/lib/python3.5/site-packages/idesolver/idesolver.py", line 257
    logger.debug(f'Advanced to iteration {self.iteration}. Current error: {error_current}.')
                                                                                          ^
SyntaxError: invalid syntax
```
while in python2 I get (installed with pip instead of pip3):
```
Python 2.7.13 (default, Dec  1 2017, 09:21:53)
[GCC 6.4.1 20170727 (Red Hat 6.4.1-1)] on linux2
>>> from idesolver import IDESolver
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File ".../.local/lib/python2.7/site-packages/idesolver/__init__.py", line 1, in <module>
    from .idesolver import *
  File ".../.local/lib/python2.7/site-packages/idesolver/idesolver.py", line 40
    def complex_quad(integrand: Callable, lower_bound: float, upper_bound: float, **kwargs) -> (complex, float, float, tuple, tuple):
                              ^
SyntaxError: invalid syntax
```
