### 2018-01-23 #542 @iljah
Here are the issues in the target repository I've created while trying out the software: https://github.com/JoshKarpel/idesolver/issues/3 https://github.com/JoshKarpel/idesolver/issues/4
