### 2018-01-23 #542 @JoshKarpel
I've started a [branch](https://github.com/JoshKarpel/idesolver/tree/joss-review) to address issues raised during the review.
