### 2018-01-23 #542 @iljah
About **installation instructions**: pip is an automatic solution but there doesn't seem to be a list of dependencies anywhere, which are scipy and numpy according to pip.
