+++ paper/paper.bib
@@ -6,7 +6,7 @@
     pages = "392 - 397",
     year = "2014",
     issn = "0010-4655",
-    doi = "https://doi.org/10.1016/j.cpc.2013.09.008",
+    doi = "10.1016/j.cpc.2013.09.008",
     url = "http://www.sciencedirect.com/science/article/pii/S0010465513003093",
     author = "Claudio A. Gelmi and Héctor Jorquera",
     keywords = "Integro-differential equation, Iterative method, Successive relaxation method, MATLAB"
