### 2018-01-23 JoshKarpel/idesolver#4 @JoshKarpel
Both of these are results of IDESolver being Python 3.6+ only, because I use f-strings. I'd like to keep this requirement unless there are strong objections.
