### 2018-01-23 JoshKarpel/idesolver#3 @JoshKarpel
I suspect you're trying to install into the system Python. If you install into a virtual environment, you shouldn't need `sudo` or `--user`. Virtual environment installation is the standard - I've added a note to this effect in the readme and the quickstart.
