### 2018-01-23 7ebb514
Added some details about installation and requirements to the readme and the quickstart.
