+++ README.rst
@@ -13,6 +13,8 @@ idesolver
 A general purpose integro-differential equation (IDE) solver, based on an iterative scheme devised by `Gelmi and Jorquera <https://doi.org/10.1016/j.cpc.2013.09.008>`_.
 
 Stable releases are available on PyPI: ``pip install idesolver``.
+IDESolver requires Python 3.6+, `numpy <https://pypi.python.org/pypi/numpy>`_, and `scipy <https://pypi.python.org/pypi/scipy/>`_.
+We recommend installing into a `virtual environment <https://docs.python.org/3/tutorial/venv.html>`_.
 
 Full documentation can be found `here <https://joshkarpel.github.io/idesolver/build/html/index.html>`_.
 
