### 2018-01-23 03bef9c
Added instructions on running the test suite.
Added some lines about the purpose of the package to the readme.
