### 2018-01-23 #542 @JoshKarpel
I've fixed the bad DOI link.

Dependencies are listed explicitly in the [requirements.txt](https://github.com/JoshKarpel/idesolver/blob/master/requirements.txt) file in the top directory of the repository, which is standard for Python/pip projects. I've added a line to the readme which lists the dependencies less formally.

I've added some information about running the test suite to the end of the manual, and linked to that in the readme. I also added a bit more information to the readme about the purpose of IDESolver.

I've responded to both issues over in the target repository.

 (All changes are in the joss-review branch, and will not appear on the hosted documentation website until they're pushed to master.)
