### 2018-01-23 JoshKarpel/idesolver#4 @arghdos
@JoshKarpel neat! I didn't know about fstrings -- I typically use a `Template` that `safe_substitute`'s `**locals` (which has it's obvious disadvantages).  You might consider adding [future-fstrings](https://github.com/asottile/future-fstrings) or the like for Python < 3.6 for backwards compatibility
