### 2018-01-24 0e3ed1c
Missed pytest-mock in the test packages
