### 2018-01-24 JoshKarpel/idesolver#4 @JoshKarpel
I spent some time today investigating backwards-compatibility options. Switching to `.format()` does allow for compatibility with 3.5, but IDESolver is still not compatible with <=3.4 because I use the `typing` module for type annotations.

I would rather not give up f-strings and type annotations, which dramatically improve the quality of the code, to gain backwards-compatibility for just a few versions. I don't know of anyone still using <=3.3 that isn't on 2.7, which I have no interest in doing compatibility for. Most users should be able to upgrade from 3.4-5 to 3.6 without trouble. Thoughts?
