### 2018-01-24 JoshKarpel/idesolver#4 @iljah
Requiring python >= 3.6 should be fine but the error message for < 3.6 users should be improved. Even better if pip refuses to install the program when running python < 3.6.
