### 2018-01-24 75790d2
Added version check to setup.py to prevent installs on <3.6
