+++ setup.py
@@ -1,5 +1,13 @@
+from __future__ import print_function
+
 from setuptools import setup, find_packages
 import os
+import sys
+
+if any((sys.version_info.major < 3, sys.version_info.major >= 3 and sys.version_info.minor < 6)):
+    version_str = '{}.{}.{}'.format(sys.version_info.major, sys.version_info.minor, sys.version_info.micro)
+    print('ERROR: IDESolver requires Python 3.6+, but you are trying to install it into Python {}'.format(version_str))
+    sys.exit(1)
 
 THIS_DIR = os.path.abspath(os.path.dirname(__file__))
 
