### 2018-01-24 JoshKarpel/idesolver#4 @JoshKarpel
> Requiring python >= 3.6 should be fine but the error message for < 3.6 users should be improved. Even better if pip refuses to install the program when running python < 3.6.

Very good idea. https://github.com/JoshKarpel/idesolver/commit/75790d21c7e96514add92abb1e2a9083f00b76b4 adds a checks to the `setup.py` that gives an error message if trying to install on Python <3.6
