### 2018-01-24 JoshKarpel/idesolver#4 @arghdos
@JoshKarpel that looks like a good compromise
