### 2018-01-25 JoshKarpel/idesolver#4 @iljah
I only tested that 2.7.14 gave a clear error, thanks.
