### 2018-01-25 JoshKarpel/idesolver#5 @iljah
__init__() got an unexpected keyword argument 'y_initial'
Don't know whether this is a documentation or code issue but trying out the quickstart:
```
import numpy as np
from idesolver import IDESolver
solver = IDESolver(
    x = np.linspace(0, 1, 100),
    y_initial = 0,
    c = lambda x, y: y - (.5 * x) + (1 / (1 + x)) - np.log(1 + x),
    d = lambda x: 1 / (np.log(2)) ** 2,
    k = lambda x, s: x / (1 + s),
    f = lambda y: y,
    lower_bound = lambda x: 0,
    upper_bound = lambda x: 1,
)
```
I get:
```
Traceback (most recent call last):
  File "<stdin>", line 9, in <module>
TypeError: __init__() got an unexpected keyword argument 'y_initial'
```
`y` instead of `y_initial` doesn't work either.
