### 2018-01-25 #542 @iljah
Another issue: https://github.com/JoshKarpel/idesolver/issues/5
