### 2018-01-25 JoshKarpel/idesolver#6 @iljah
Test failures
Running pytest gives me:
```
================================================================= ERRORS ==================================================================
____________________________________ ERROR at setup of test_callback_is_called_correct_number_of_times ____________________________________
file .../idesolver/tests/test_misc.py, line 73
  def test_callback_is_called_correct_number_of_times(mocker):
E       fixture 'mocker' not found
>       available fixtures: cache, capfd, capfdbinary, caplog, capsys, capsysbinary, default_solver, doctest_namespace, dummy_args, monkeypatch, pytestconfig, record_xml_property, recwarn, tmpdir, tmpdir_factory
>       use 'pytest --fixtures [testpath]' for help on them.

.../idesolver/tests/test_misc.py:73
================================================================ FAILURES =================================================================
______________________________________________ test_raise_exception_if_unexpectedly_complex _______________________________________________
tests/test_complex_valued.py:21: in test_raise_exception_if_unexpectedly_complex
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
__________________________________________________ test_no_exception_if_expected_complex __________________________________________________
tests/test_complex_valued.py:36: in test_no_exception_if_expected_complex
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
_________________________________________________ test_warning_when_not_enough_iterations _________________________________________________
tests/test_misc.py:26: in test_warning_when_not_enough_iterations
    good_solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
________________________________ test_number_of_intermediate_solutions_is_same_as_iteration_count_plus_one ________________________________
tests/test_misc.py:67: in test_number_of_intermediate_solutions_is_same_as_iteration_count_plus_one
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
_________________________________________ test_real_ide_against_analytic_solution[solver0-exact0] _________________________________________
tests/test_solver_against_analytic_solutions.py:114: in test_real_ide_against_analytic_solution
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
_________________________________________ test_real_ide_against_analytic_solution[solver1-exact1] _________________________________________
tests/test_solver_against_analytic_solutions.py:114: in test_real_ide_against_analytic_solution
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
_________________________________________ test_real_ide_against_analytic_solution[solver2-exact2] _________________________________________
tests/test_solver_against_analytic_solutions.py:114: in test_real_ide_against_analytic_solution
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
_________________________________________ test_real_ide_against_analytic_solution[solver3-exact3] _________________________________________
tests/test_solver_against_analytic_solutions.py:114: in test_real_ide_against_analytic_solution
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
_________________________________________ test_real_ide_against_analytic_solution[solver4-exact4] _________________________________________
tests/test_solver_against_analytic_solutions.py:114: in test_real_ide_against_analytic_solution
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
_________________________________________ test_real_ide_against_analytic_solution[solver5-exact5] _________________________________________
tests/test_solver_against_analytic_solutions.py:114: in test_real_ide_against_analytic_solution
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
_________________________________________ test_real_ide_against_analytic_solution[solver6-exact6] _________________________________________
tests/test_solver_against_analytic_solutions.py:114: in test_real_ide_against_analytic_solution
    solver.solve()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:249: in solve
    y_current = self._initial_y()
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:296: in _initial_y
    return self._solve_ode(self.c)
../../.local/lib/python3.6/site-packages/idesolver/idesolver.py:363: in _solve_ode
    sol = integ.solve_ivp(
E   AttributeError: module 'scipy.integrate' has no attribute 'solve_ivp'
============================================== 11 failed, 64 passed, 1 error in 2.18 seconds ==============================================```
