### 2018-01-25 #542 @iljah
There might be missing dependencies in the manual regarding testing: https://github.com/JoshKarpel/idesolver/issues/6
