### 2018-01-25 51f92c1
Adds minimum versions of numpy and scipy to setup.py
