+++ requirements.txt
@@ -1,2 +1,2 @@
-numpy==1.13.3
-scipy==1.0.0
+numpy>=1.13.3
+scipy>=1.0.0
