### 2018-01-25 JoshKarpel/idesolver#6 @JoshKarpel
I suspect this is happening because your `scipy` is not at least `1.0.0` - I checked my `setup.py` and realized that I never actually specified a version for either `numpy` or `scipy`, only in my `requirements.txt`.

https://github.com/JoshKarpel/idesolver/commit/51f92c1617be9f996082794a8f52c625b641c75c fixes this by adding minimum version requirements to `setup.py`. This won't be reflected on PyPi yet - you'll need to update `scipy` manually.
