### 2018-01-25 045451c
Fixed doc typo (y_initial -> y_0)
