### 2018-01-25 JoshKarpel/idesolver#5 @JoshKarpel
This was a docs typo. Fixed in https://github.com/JoshKarpel/idesolver/commit/045451cba9db36c762d223d9374c7cbaf276aa22
