### 2018-01-25 JoshKarpel/idesolver#3 @JoshKarpel
@iljah - could you clarify whether the above comment fixed the issue so that I can close?
