### 2018-01-26 #542 @JoshKarpel
Question for @kyleniemeyer - some of the changes need to be pushed to PyPi as a new version, so the post-review version will actually be 1.0.2, not 1.0.1 (hopefully; assuming no breaking changes are needed). Is it possible to change the version number that goes into the final documents?
