### 2018-01-26 #542 @kyleniemeyer
@JoshKarpel don't worry about the version number, it's expected that it will change as a result of reviewer-recommended changes.
