### 2018-01-26 JoshKarpel/idesolver#5 @iljah
y_0 = 0 works, thanks.
