### 2018-01-26 JoshKarpel/idesolver#6 @iljah
Upgrading scipy left one error, looks like a missing dependency:
```
================================================================= ERRORS =============================
____________________________________ ERROR at setup of test_callback_is_called_correct_number_of_times ________
file .../idesolver/tests/test_misc.py, line 73
  def test_callback_is_called_correct_number_of_times(mocker):
E       fixture 'mocker' not found
>       available fixtures: cache, capfd, capfdbinary, caplog, capsys, capsysbinary, default_solver, doctest_namespace, dummy_args, monkeypatch, pytestconfig, record_xml_property, recwarn, tmpdir, tmpdir_factory
>       use 'pytest --fixtures [testpath]' for help on them.

.../idesolver/tests/test_misc.py:73
=================================================== 75 passed, 1 error in 63.55 seconds ===```
