### 2018-01-26 JoshKarpel/idesolver#3 @iljah
Virtual environment also works but in that case it would be helpful to show all required commands, which I guess could be:
```
python3 -m venv $HOME/idesolver-env
pip install idesolver
```
assuming the user already has pip installed, etc.
