### 2018-01-26 JoshKarpel/idesolver#6 @JoshKarpel
Oh, damn, I left that one out of the list initially. The missing dependency is `pytest-mock`. I added it to the list in https://github.com/JoshKarpel/idesolver/commit/0e3ed1c6da8a28ca7bbe87d3ed0d25eaa30a1592.
