### 2018-01-26 JoshKarpel/idesolver#3 @JoshKarpel
I'd prefer not to explicitly write out the instructions, instead linking to some external resource (which was done in https://github.com/JoshKarpel/idesolver/commit/7ebb514ce1f958d8ef9cf8f23c5830ffb5953dfc). Although installing into virtual environments is the standard, the preferred method of doing that is changing pretty fast these days (I use conda, for example, and pipenv is becoming popular recently). I like the current solution of linking to the official docs on virtual environments, since it's the common denominator where any virtual environment solution starts from. The instructions for any particular solution are going to be quite different from each other, and any specifics given may conflict with whatever the user already has installed.
