### 2018-01-29 JoshKarpel/idesolver#6 @iljah
Thanks, all test pass now.
