### 2018-01-29 #542 @iljah
All issues have been resolved so this looks good to go.
