### 2018-01-30 e0654c1
Ensure HTML docs are updated
