+++ docs/build/html/_sources/manual.rst.txt
@@ -73,6 +73,6 @@ A different global error estimator can be passed in the constructor as the argum
 Test Suite
 ----------
 
-Running the test suite requires some additional Python packages: `pip install pytest hypothesis`.
+Running the test suite requires some additional Python packages: `pip install pytest pytest-mock hypothesis`.
 You'll also need to get the entire IDESolver repository via `git clone https://github.com/JoshKarpel/idesolver.git`.
 Once installed, you can run the test suite by running `pytest` from the top-level directory of the package.
