+++ docs/source/changelog.rst
@@ -3,6 +3,10 @@ Change Log
 
 .. currentmodule:: idesolver
 
+v1.0.2
+------
+* IDESolver now explicitly requires Python 3.6+ on install. Dependencies on ``numpy`` and ``scipy`` are given as lower bounds.
+
 v1.0.1
 ------
 * Changed the name of ``IDESolver.F`` to ``f``.
