+++ setup.py
@@ -16,7 +16,7 @@ with open(os.path.join(THIS_DIR, 'README.rst')) as f:
 
 setup(
     name = 'idesolver',
-    version = '1.0.1',
+    version = '1.0.2',
     author = 'Josh Karpel',
     author_email = 'karpel@wisc.edu',
     maintainer = 'Josh Karpel',
