+++ docs/source/conf.py
@@ -56,7 +56,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'idesolver'
-copyright = '2017, Joshua T Karpel'
+copyright = '2017-2018, Joshua T Karpel'
 author = 'Joshua T Karpel'
 
 # The version info for the project you're documenting, acts as replacement for
@@ -64,9 +64,9 @@ author = 'Joshua T Karpel'
 # built documents.
 #
 # The short X.Y version.
-version = '1.0'
+version = '1.0.2'
 # The full version, including alpha/beta/rc tags.
-release = '1.0'
+release = '1.0.2'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
