### 2018-01-31 #542 @JoshKarpel
I've merged the responses to the review issues (they were very helpful, thanks @iljah and @arghdos!), and released v1.0.2.

@kyleniemeyer I got a Zenodo: DOI https://doi.org/10.5281/zenodo.1162964
