### 2018-01-31 #542 @kyleniemeyer
Hi @JoshKarpel—looks good (including the paper)!
