### 2018-01-31 #542 @arfon
@iljah @arghdos - many thanks for your reviews here and to @kyleniemeyer for editing this submission ✨

@JoshKarpel - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00542 ⚡️ 🚀 :boom:
